Modified the State class
Added an overload to the populate method that takes a player object as an argument.

Changed the width and height to 800 by 600.

Modified the main class to it calls the populate method each time a player has successfully logged in. And a player will only keep sending moves if it's still alive.